Misc (re #1945): Added feature of listing dshow device in ffmpeg video device (experimental).
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@5441 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip/src/pjsua2/endpoint.cpp
@@ -490,7 +490,7 @@ void Endpoint::logFunc(int level, const char *data, int len)
     LogEntry entry;
     entry.level = level;
     entry.msg = string(data, len);
-    entry.threadId = (long)pj_thread_this();
+    entry.threadId = (long)(size_t)pj_thread_this();
     entry.threadName = string(pj_thread_get_name(pj_thread_this()));
 
     ep.utilLogWrite(entry);
